Fix runtime loading sitelet from value [<Website>]

+++ tools/WebSharper.Fake.fsx
@@ -301,27 +301,28 @@ let MakeTargets (args: Args) =
         let re = Regex(@"^(\s*(\S+)\s*~>)\s*LOCKEDVERSION/([1-3])")
         let lock = Paket.LockFile.LoadFrom "paket.lock"
         let g = lock.Groups.[Paket.Domain.GroupName "main"]
-        for f in Directory.EnumerateFiles("nuget", "*.paket.template.in") do
-            let s =
-                File.ReadAllLines(f)
-                |> Array.map (fun l ->
-                    re.Replace(l, fun m ->
-                        let init = m.Groups.[1].Value
-                        let pkg = m.Groups.[2].Value
-                        let prefixLen = int m.Groups.[3].Value
-                        let v = g.GetPackage(Paket.Domain.PackageName pkg).Resolved.Version
-                        let pre =
-                            match v.PreRelease with
-                            | None -> ""
-                            | Some x -> "-" + x.Name
-                        match prefixLen with
-                        | 1 -> sprintf "%s %i%s" init v.Major pre
-                        | 2 -> sprintf "%s %i.%i%s" init v.Major v.Minor pre
-                        | 3 -> sprintf "%s %i.%i.%i%s" init v.Major v.Minor v.Patch pre
-                        | _ -> failwith "Impossible"))
-            let outName = f.[..f.Length-4]
-            printfn "Writing %s" outName
-            File.WriteAllLines(outName, s)
+        if Directory.Exists("nuget") then
+            for f in Directory.EnumerateFiles("nuget", "*.paket.template.in") do
+                let s =
+                    File.ReadAllLines(f)
+                    |> Array.map (fun l ->
+                        re.Replace(l, fun m ->
+                            let init = m.Groups.[1].Value
+                            let pkg = m.Groups.[2].Value
+                            let prefixLen = int m.Groups.[3].Value
+                            let v = g.GetPackage(Paket.Domain.PackageName pkg).Resolved.Version
+                            let pre =
+                                match v.PreRelease with
+                                | None -> ""
+                                | Some x -> "-" + x.Name
+                            match prefixLen with
+                            | 1 -> sprintf "%s %i%s" init v.Major pre
+                            | 2 -> sprintf "%s %i.%i%s" init v.Major v.Minor pre
+                            | 3 -> sprintf "%s %i.%i.%i%s" init v.Major v.Minor v.Patch pre
+                            | _ -> failwith "Impossible"))
+                let outName = f.[..f.Length-4]
+                printfn "Writing %s" outName
+                File.WriteAllLines(outName, s)
         Paket.Pack <| fun p ->
             { p with
                 OutputPath = "build"
